requirements.txt
requests
beautifulsoup4
googletrans
selenium
webdriver-manager
python-dotenv       
pandas
